fix: handle clicks
