+++ frontend/src/scenes/max/HistoryPreview.tsx
@@ -19,7 +19,7 @@ export function HistoryPreview({ sidePanel = false }: HistoryPreviewProps): JSX.
     }
 
     return (
-        <div className="w-full max-w-120 flex flex-col gap-2">
+        <div className="max-w-120 w-full self-center flex flex-col gap-2">
             <div className="flex items-center justify-between gap-2 -mr-2">
                 <h3 className="text-sm font-medium text-secondary mb-0">Recent chats</h3>
                 <LemonButton
